requirements.txt
Click==7.0
ipdb==0.11
jupyter==1.0.0
matplotlib==2.2.3
numpy==1.15.2
pandas==0.23.4
pendulum==2.0.4
pytest==4.1.1
PyYAML==3.13
sklearn==0.0
zope
ipython
